Refactor application to achieve a zero-warning build
Resolve all 40 compiler and lint warnings while preserving the
application's existing architecture and behavior. The refactoring
removes unused components and repository functions, simplifies
redundant logic, and applies modern Kotlin conventions consistently.

Use Kotlin Sequences where collection processing benefits from lazy
evaluation, remove unnecessary parentheses and implicit lambda
parameter shadowing, and add trailing commas throughout the modified
source and test files. Standardize braces to follow Allman style so the
code remains consistent with the project's formatting requirements.

Remove deprecated system bar configuration and unused dynamic color
logic from Theme.kt because enableEdgeToEdge() already provides the
required system UI behavior. This avoids maintaining redundant,
deprecated API usage and eliminates related warnings.

Correct BigDecimal assertion handling in the checkout tests so
delta-based comparisons use the expected Double values. Refine the
mock listener logic in AuthRepositoryTest.kt to remove unchecked cast
warnings without changing the tested repository behavior.

The resulting build reports zero errors and zero warnings from
:app:compileDebugKotlin. Gradle synchronization and the affected unit
tests were also verified to ensure that the cleanup did not introduce
regressions.

+++ app/src/test/java/com/example/campus_eats_app_kt/util/CheckoutEngineTest.kt
@@ -30,11 +30,11 @@ class CheckoutEngineTest
         // Discount (0%) = 0.0
         // Raw Total = 260.0
         // Rounding to R5 = 260.0
-        assertEquals(200.0, result.subtotal, 0.001)
-        assertEquals(40.0, result.tax, 0.001)
-        assertEquals(20.0, result.serviceFee, 0.001)
-        assertEquals(0.0, result.studentDiscount, 0.001)
-        assertEquals(260.0, result.total, 0.001)
+        assertEquals(200.0, result.subtotal.toDouble(), 0.001)
+        assertEquals(40.0, result.tax.toDouble(), 0.001)
+        assertEquals(20.0, result.serviceFee.toDouble(), 0.001)
+        assertEquals(0.0, result.studentDiscount.toDouble(), 0.001)
+        assertEquals(260.0, result.total.toDouble(), 0.001)
     }
 
     /**
@@ -57,10 +57,10 @@ class CheckoutEngineTest
         // Discount (0%) = 0.0
         // Raw Total = 759.0
         // Rounding to next R5 = 760.0
-        assertEquals(600.0, result.subtotal, 0.001)
-        assertEquals(120.0, result.tax, 0.001)
-        assertEquals(39.0, result.serviceFee, 0.001)
-        assertEquals(760.0, result.total, 0.001)
+        assertEquals(600.0, result.subtotal.toDouble(), 0.001)
+        assertEquals(120.0, result.tax.toDouble(), 0.001)
+        assertEquals(39.0, result.serviceFee.toDouble(), 0.001)
+        assertEquals(760.0, result.total.toDouble(), 0.001)
     }
 
     /**
@@ -82,8 +82,8 @@ class CheckoutEngineTest
         // Service Fee (0%) = 0.0
         // Raw Total = 1440.0
         // Rounding to R5 = 1440.0
-        assertEquals(0.0, result.serviceFee, 0.001)
-        assertEquals(1440.0, result.total, 0.001)
+        assertEquals(0.0, result.serviceFee.toDouble(), 0.001)
+        assertEquals(1440.0, result.total.toDouble(), 0.001)
     }
 
     /**
@@ -94,11 +94,11 @@ class CheckoutEngineTest
     {
         // Boundary R500 (inclusive of mid tier)
         val res500 = CheckoutEngine.calculateSummary(500.0, UserRole.STANDARD)
-        assertEquals(500.0 * 0.065, res500.serviceFee, 0.001)
+        assertEquals(500.0 * 0.065, res500.serviceFee.toDouble(), 0.001)
 
         // Boundary R1000 (inclusive of mid tier)
         val res1000 = CheckoutEngine.calculateSummary(1000.0, UserRole.STANDARD)
-        assertEquals(1000.0 * 0.065, res1000.serviceFee, 0.001)
+        assertEquals(1000.0 * 0.065, res1000.serviceFee.toDouble(), 0.001)
     }
 
     /**
@@ -121,8 +121,8 @@ class CheckoutEngineTest
         // Student Discount (2.5%) = 10.0
         // Raw Total = 400 + 80 + 40 - 10 = 510.0
         // Rounding to R5 = 510.0
-        assertEquals(10.0, result.studentDiscount, 0.001)
-        assertEquals(510.0, result.total, 0.001)
+        assertEquals(10.0, result.studentDiscount.toDouble(), 0.001)
+        assertEquals(510.0, result.total.toDouble(), 0.001)
     }
 
     /**
@@ -146,6 +146,6 @@ class CheckoutEngineTest
         val result = CheckoutEngine.calculateSummary(subtotal, role)
 
         // Then:
-        assertEquals(135.0, result.total, 0.001)
+        assertEquals(135.0, result.total.toDouble(), 0.001)
     }
 }
